moved old tutorial code to a folder and reformatted app.py code to be reused in future.
